rag-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
google-generativeai==0.3.1
faiss-cpu==1.7.4
rank_bm25==0.2.2
numpy==1.26.2
python-decouple==3.8
